requirements.txt
qrcode
Pillow
tkinter
